% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{addCodesetAnnotation}
\alias{addCodesetAnnotation}
\title{Add NanoString codeset annotation to a NanoString ExpressionSet object}
\usage{
addCodesetAnnotation(rccSet, annot, reorder = TRUE, showWarnings = TRUE)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object.}

\item{annot}{Data frame containing the codeset annotation.}

\item{reorder}{Logical indicating whether the probes should be reordered
according to their barcodes (this can help in identifying
barcode-specific artifacts -- i.e. background noise).}

\item{showWarnings}{Logical indicating whether or not warnings should be shown, if any.}
}
\value{
A copy of the input ExpressionSet where the codeset annotation has been
merged into its fData slot.
}
\description{
Returns a copy of the input ExpressionSet where the codeset annotation
                has been merged into its fData slot. The merge key for each is a string
                formed from the concatenation of their CodeClass, GeneName, and
                Accession columns ("<CodeClass>_<GeneName>_<Accession>"). For creating
                the codeset annotation object, see buildCodesetAnnotation().
}
\examples{
rccDir <- system.file("extdata", "RCC", package="NanoStringQCPro")
rccSet <- readRccBatch(rccDir)
rlfPath <- system.file("extdata", "RLF", "NQCP_example.rlf", package="NanoStringQCPro")
annot <- buildCodesetAnnotation(rlfPath)
rccSet.annotated <- addCodesetAnnotation(rccSet, annot)
}
\author{
Dorothee Nickles, Robert Ziman
}

