% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\docType{methods}
\name{addQCFlags,RccSet-method}
\alias{addQCFlags}
\alias{addQCFlags,RccSet-method}
\title{Add sample QC flags to an rccSet}
\usage{
\S4method{addQCFlags}{RccSet}(rccSet, flagFile)
}
\arguments{
\item{rccSet}{An RccSet object}

\item{flagFile}{Path to a sample QC flag file as generated by the
                    NanoStringQCPro QC report (see makeQCReport())}
}
\value{
A copy of the input RccSet with columns added to pData from the QC
flag file.
}
\description{
Returns a copy of the input RccSet with columns added to pData
                from the provided sample QC flag annotation file. (That file is
                produced by makeQCReport(); see its help page for more
                details.)
}
\author{
Dorothee Nickles
}

