% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{bdPlot,RccSet-method}
\alias{bdPlot}
\alias{bdPlot,RccSet-method}
\title{Binding density plot}
\usage{
\S4method{bdPlot}{RccSet}(rccSet)
}
\arguments{
\item{rccSet}{An RccSet object}
}
\value{
A plot
}
\description{
Plot the binding density of each sample in an RccSet object.
Samples with a binding density < 0.05 or > 2.25 (thresholds defined by NanoString)
are marked in red (dashed red line indicates threshold).
}
\author{
Dorothee Nickles
}

