% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{checkRccSet,RccSet-method}
\alias{checkRccSet}
\alias{checkRccSet,RccSet-method}
\title{Check an RccSet}
\usage{
\S4method{checkRccSet}{RccSet}(rccSet, reportWarnings = TRUE,
  showMessages = FALSE)
}
\arguments{
\item{rccSet}{An RccSet to be checked.}

\item{reportWarnings}{Logical. If TRUE, warnings are reported.}

\item{showMessages}{Logical. If TRUE, notes are shown indicating any optional missing
                        columns and the like.}
}
\value{
Returns TRUE if no warnings were generated and FALSE otherwise.
}
\description{
Provides additional checks and generates warnings for unexpected or unusual conditions
which, though permitted by the RccSet class, may indicate data import errors.
}
\examples{
data(example_rccSet)
checkRccSet(example_rccSet)
}
\author{
Robert Ziman
}

