% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{colByFun}
\alias{colByFun}
\title{colByFun}
\usage{
colByFun(x, thresholds)
}
\arguments{
\item{x}{Numeric vector}

\item{thresholds}{List of length 2, with a scalar numeric in each slot, one giving the lower the upper threshold (for outlier definition)}
}
\value{
A vector of colors, with "red" for all values of x exceeding thresholds and "black" for all other values
}
\description{
Color x based on upper and lower thresholds
}
\author{
Dorothee Nickles
}

