% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{countsInBlankSamples_verticalPlot}
\alias{countsInBlankSamples_verticalPlot}
\title{Plot counts in blank samples (vertical orientation)}
\usage{
countsInBlankSamples_verticalPlot(rccSet, outputFile)
}
\arguments{
\item{rccSet}{An RccSet}

\item{outputFile}{Output PNG filename}
}
\value{
A PNG file containing a boxplot of the gene-wise counts for the blank samples
in the input. The PNG is set to a fixed resolution of 300 pixels per inch and
a fixed width of 2250 pixels (i.e. 7.5" at 300ppi), but the height varies
with the size of the input. The font size is also fixed so that the labels
will be legible even for large datasets.
}
\description{
Plot counts in blank samples (vertical orientation)
}

