% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{ctrlsOverviewPlot,RccSet-method}
\alias{ctrlsOverviewPlot}
\alias{ctrlsOverviewPlot,RccSet-method}
\title{ctrlsOverviewPlot}
\usage{
\S4method{ctrlsOverviewPlot}{RccSet}(rccSet)
}
\arguments{
\item{rccSet}{An RccSet object}
}
\value{
A plot with two panels, one for the negative controls, one for the positive controls.
}
\description{
Plot individual negative and positive controls across all samples in an RccSet object.
}
\author{
Dorothee Nickles
}

