% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{ctrlsPlot}
\alias{ctrlsPlot}
\title{ctrlsPlot}
\usage{
ctrlsPlot(rccSet, method = c("cutoffByMMAD", "cutoffByVar"), cutoff)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{method}{scalar character, specifying the method for outlier detection; either cutoffByMMAD or cutoffByVar}

\item{cutoff}{scalar numeric, cutoff in method to determine outliers}
}
\value{
Plots generated by ctrlsOverviewPlot, iqrPlot, posSlopePlot, posMaxPlot, allSumPlot and sumRatioPlot.
}
\description{
Wrapper function for plotting QC metrics involving NanoString controls
}
\author{
Dorothee Nickles
}
\seealso{
\code{\link{ctrlsOverviewPlot}}, \code{\link{iqrPlot}}, \code{\link{posSlopePlot}}, \code{\link{posMaxPlot}},
\code{\link{allSumPlot}}, \code{\link{sumRatioPlot}}
}

