% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{ctrlsZprimePlot,RccSet-method}
\alias{ctrlsZprimePlot,RccSet-method}
\alias{ctrlsZprimePlot}
\title{ctrlsZprimePlot}
\usage{
\S4method{ctrlsZprimePlot}{RccSet}(rccSet)
}
\arguments{
\item{rccSet}{An RccSet object}
}
\value{
A plot
}
\description{
Plot distribution of counts and the Z' Factors comparing the nagative controls and
                the three highest input positive controls of an RccSet object.
}
\author{
Dorothee Nickles
}
