% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{cutoffByMMAD}
\alias{cutoffByMMAD}
\title{cutoffByMMAD}
\usage{
cutoffByMMAD(x, d, ...)
}
\arguments{
\item{x}{numeric vector}

\item{d}{scalar numeric, factor by which to multiply MAD of x}

\item{...}{additional parameters passed on to median()}
}
\value{
A list of length 2, with a scalar numeric in each slot, one giving the lower threshold (median(x) - d * mad(x)),
the other giving the upper threshold (median(x) + d * mad(x)) for outlier definition.
}
\description{
Determine cutoffs of x (for outlier detection) based on median
}
\author{
Dorothee Nickles
}

