% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{cutoffByVar}
\alias{cutoffByVar}
\title{cutoffByVar}
\usage{
cutoffByVar(x, d, ...)
}
\arguments{
\item{x}{numeric vector}

\item{d}{scalar numeric, percent CV; passed on to dCoVar}

\item{...}{additional parameters passed on to mean()}
}
\value{
A list of length 2, with a scalar numeric in each slot, one giving the lower threshold (mean(x) - CV% based cutoff),
the other giving the upper threshold (mean(x) + percent CV based cutoff) for outlier definition.
}
\description{
Determine cutoffs of x (for outlier detection) based on a certain percent CV
}
\author{
Dorothee Nickles
}

