% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{dCoVar}
\alias{dCoVar}
\title{dCoVar}
\usage{
dCoVar(x, d, ...)
}
\arguments{
\item{x}{numeric vector}

\item{d}{scalar numeric, percent CV}

\item{...}{additional parameters passed on to mean()}
}
\value{
standard deviation of x at d percent of CV
}
\description{
Determine standard deviation at a certain percent CV
}
\author{
Dorothee Nickles
}

