% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{densityPlot}
\alias{densityPlot}
\title{densityPlot}
\usage{
densityPlot(M, log.transform = FALSE, pdata, covar, ...)
}
\arguments{
\item{M}{One of the matrices from the assayData() of an RccSet object (make sure
to set the log.transform parameter accordingly)}

\item{log.transform}{Scalar boolean}

\item{pdata}{pData() of the RccSet object}

\item{covar}{character; colname in pData() that can be used to label genes by a category of interest (passed on to colByCovar)}

\item{...}{additional plotting parameters}
}
\value{
A density plot
}
\description{
Plot the density of counts for all endogenous and housekeeping genes for each sample in an RccSet object
}
\author{
Dorothee Nickles
}

