% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example_rccSet.R
\docType{data}
\name{example_rccSet}
\alias{example_rccSet}
\title{NanoStringQCPro example dataset}
\format{A NanoString ExpressionSet object}
\source{
This is an artificial dataset designed to resemble real data.
}
\description{
Example data for the NanoStringQCPro package
}
\author{
Dorothee Nickles
}
\keyword{datasets}

