% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{flagSamplesTech,RccSet-method}
\alias{flagSamplesTech,RccSet-method}
\alias{flagSamplesTech}
\title{flagSamplesTech}
\usage{
\S4method{flagSamplesTech}{RccSet}(rccSet)
}
\arguments{
\item{rccSet}{An RccSet object}
}
\value{
A numeric vector giving the indices of samples with outlier values in FOV counted and
binding density < 0.05 or > 2.25 (thresholds defined by NanoString) will be flagged.
}
\description{
Flag samples based on their technical performance, i.e. field of vision (FOV) counted and binding density
}
\details{
Samples with a FOV counted/FOV count of less than 80% will be flagged. Also samples with a binding density
}
\author{
Dorothee Nickles
}
