% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{fovPlot}
\alias{fovPlot}
\title{fovPlot}
\usage{
fovPlot(rccSet)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}
}
\value{
A plot
}
\description{
Plot the fraction of successfully imaged fields of view (FOV) in the given NanoString ExpressionSet.
The ExpressionSet's pData should have 'FovCount' and 'FovCounted' columns populated with the total
and successfully imaged FOV counts, respectively. Samples with a FOV counted/FOV count of less than
80% are marked in red (dashed red line indicates threshold).
}
\author{
Dorothee Nickles
}

