% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{geneClustering}
\alias{geneClustering}
\title{Gene clustering heatmap}
\usage{
geneClustering(rccSet, outputFile, main = "Gene clustering", covar = NULL)
}
\arguments{
\item{rccSet}{An RccSet}

\item{outputFile}{Output PNG filename}

\item{main}{Plot title}

\item{covar}{Colname in the rccSet's fData that can be used to label genes by a category of interest}
}
\value{
A PNG file showing clustering of genes by correlation across an experiment
Positive and negative control probes and any zero-variance genes (typically
housekeeping genes) are omitted from the heatmap.
The width and height of the PNG file are set to vary with the size of the input.
}
\description{
Gene clustering heatmap
}
\author{
Dorothee Nickles, Robert Ziman
}

