% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{geneClusteringPNG}
\alias{geneClusteringPNG}
\title{geneClustering}
\usage{
geneClusteringPNG(M, main, labels, mainImage, previewImage,
  log.transform = FALSE, covar = NULL, fdata = NULL, gsubset = NULL,
  ...)
}
\arguments{
\item{M}{One of the matrices from assayData(rccSet) where rccSet is a NanoString ExpressionSet object}

\item{main}{Plot title}

\item{labels}{Character vector for the axis labels}

\item{mainImage}{Filename for the main PNG file that will be generated}

\item{previewImage}{Filename for the preview PNG file that will be generated}

\item{log.transform}{Scalar boolean}

\item{covar}{Character; colname in fdata that can be used to label genes by a category of interest}

\item{fdata}{fData(rccSet)}

\item{gsubset}{Either a boolean vector of length nrow(exprs(rccSet)) or a numeric vector of indices, indicating
to which genes in exprs(rccSet) to subset}

\item{...}{Extra arguments passed to aheatmap() (optional)}
}
\value{
A NanoString ExpressionSet object that has count data adjusted by positive control counts.
}
\description{
Clustering of genes by correlation across an experiment
}
\author{
Dorothee Nickles, Robert Ziman
}

