% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{iqrPlot}
\alias{iqrPlot}
\title{iqrPlot}
\usage{
iqrPlot(rccSet, codeClass = c("Negative", "Positive", "Endogenous",
  "Housekeeping"), method = c("cutoffByMMAD", "cutoffByVar"), cutoff)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{codeClass}{Scalar character, specifying the code class (as annotated in the fData(rccSet)$CodeClass column)
for which the IQR shall be determined}

\item{method}{Scalar character, specifying the method for outlier detection; either cutoffByMMAD or cutoffByVar}

\item{cutoff}{Scalar numeric, cutoff in method to determine outliers}
}
\value{
A plot
}
\description{
Plot the interquartile range (IQR) for a certain code class of probes in a NanoString ExpressionSet object.
}
\details{
IQR of the specified code class for each sample in the ExpressionSet are plotted and outliers (as determined
by method) are marked in red (thresholds for outlier definition are plotted as red dashed lines).
}
\author{
Dorothee Nickles
}
\seealso{
\code{\link{cutoffByMMAD}}, \code{\link{cutoffByVar}}
}

