% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{lodAssess}
\alias{lodAssess}
\title{lodAssess}
\usage{
lodAssess(rccSet, reference = c("blank", "negatives"), stringency = 2.5)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{reference}{scalar character, reference to determine limit of detection; either "blank" or "negatives"}

\item{stringency}{factor by which deviation (MAD) of the median of the reference counts is multiplied}
}
\value{
A numeric vector giving the number of missing genes (endogenous and housekeeping genes) for each sample
in a NanoString ExpressionSet.
}
\description{
Assess how many genes per sample in a NanoString ExpressionSet object are below the limit of detection.
}
\author{
Dorothee Nickles
}

