% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{lodPlot,RccSet-method}
\alias{lodPlot}
\alias{lodPlot,RccSet-method}
\title{lodPlot}
\usage{
\S4method{lodPlot}{RccSet}(rccSet, maxMiss = 0.2)
}
\arguments{
\item{rccSet}{An RccSet object}

\item{maxMiss}{Numeric specifying the allowable fraction of genes below the lower limit
of detection in a sample.}
}
\value{
A plot
}
\description{
Function to plot the number of missing genes per sample in an RccSet object.
}
\details{
Samples with more than 50% missingness are marked in red. If blank
measurements are present, they are represented as triangles.
}
\author{
Dorothee Nickles
}

