% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{negCtrlsByLane}
\alias{negCtrlsByLane}
\title{negCtrlsByLane}
\usage{
negCtrlsByLane(rccSet)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}
}
\value{
A plot with boxplots for the negative control counts for each individual sample (lane-specific background)
}
\description{
Plot negative controls per lane in a NanoString ExpressionSet object
}
\details{
Boxplots are colored by lane (as specified in the pData slot). Bars on top of
the panel indicate the stage position for each cartridge/sample (as specified in the pData slot).
}
\author{
Dorothee Nickles
}

