% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{negCtrlsPairs,RccSet-method}
\alias{negCtrlsPairs}
\alias{negCtrlsPairs,RccSet-method}
\title{negCtrlsPairs}
\usage{
\S4method{negCtrlsPairs}{RccSet}(rccSet, log.transform = FALSE)
}
\arguments{
\item{rccSet}{An RccSet object}

\item{log.transform}{boolean, whether data needs to be log2 transformed}
}
\value{
Pairs plot of the negative controls with a scatter plot in the lower panel and
correlatation coefficients printed in the upper panel.
}
\description{
Pairs plot of negative controls across all samples in an RccSet object
}
\author{
Dorothee Nickles
}

