% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{negCtrlsPlot,RccSet-method}
\alias{negCtrlsPlot}
\alias{negCtrlsPlot,RccSet-method}
\title{negCtrlsPlot}
\usage{
\S4method{negCtrlsPlot}{RccSet}(rccSet)
}
\arguments{
\item{rccSet}{An RccSet object}
}
\value{
A plot with two panels: one showing boxplots for the individual negative controls across all samples, and
one showing boxplots for the negative control counts for each individual sample (lane-specific background).
}
\description{
Plot negative controls across all samples in an RccSet object
}
\details{
In the second panel, boxplots are colored by lane (as specified in the pData slot). Bars on top of
the panel indicate the stage position for each cartridge/sample (as specified in the pData slot).
}
\author{
Dorothee Nickles
}

