% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{nSolverCsv.to.pdata_fdata_adata}
\alias{nSolverCsv.to.pdata_fdata_adata}
\title{nSolverCsv.to.pdata_fdata_fdata}
\usage{
nSolverCsv.to.pdata_fdata_adata(rccCollectorToolExport)
}
\arguments{
\item{rccCollectorToolExport}{Path to the nSolver RCC Collector Tool .CSV export.}
}
\value{
A list containing matrices (for pdata and adata) and a data frame (for fdata)
that pdata_fdata_adata.to.rccSet() then transforms into a full ExpessionSet.
}
\description{
First stage of readRccCollectorToolExport(): produces a list containing
                matrices (for pdata and adata) and a data frame (for fdata) that
                pdata_fdata_adata.to.rccSet then transforms into a full RccSet
                (after some further checks and adjustments). Not intended for
                external use; see also rccFiles.to.pdata_fdata_adata().
}
\author{
Dorothee Nickles, Thomas Sandmann, Robert Ziman
}

