% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{pdata_fdata_adata.to.rccSet}
\alias{pdata_fdata_adata.to.rccSet}
\title{pdata_fdata_adata.to.rccSet}
\usage{
pdata_fdata_adata.to.rccSet(pdata_fdata_adata)
}
\arguments{
\item{pdata_fdata_adata}{List containing the pdata, fdata, and adata returned by
                            rccFiles.to.pdata_fdata_adata() or nSolverCsv.to.pdata_fdata_adata().}
}
\value{
An \code{\linkS4class{RccSet}} whose contents reflect the input data.
}
\description{
Second stage of readRccBatch()/readRccCollectorToolExport() -- not intended for external use.
}
\details{
Note that a pseudo-count of 1 is always added to all measurements, to enable
subsequent log transformation of the data in cases where zero-counts are
present.

N.B. The function currently expects certain columns to be present in
\code{pdata_fdata_a_data$pdata.m}, and it converts these to numerics. These
expectations should be incorporated into the class definition, and
conversion should only take place with a warning. Future updates will
address this.
}
\author{
Robert Ziman
}

