% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{posCtrlsPlot}
\alias{posCtrlsPlot}
\title{posCtrlsPlot}
\usage{
posCtrlsPlot(rccSet, method = c("cutoffByMMAD", "cutoffByVar"), cutoff)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{method}{Scalar character specifying the method for outlier detection; either cutoffByMMAD or cutoffByVar}

\item{cutoff}{Scalar numeric: cutoff in method to determine outliers}
}
\value{
A plot with two panels
}
\description{
Wrapper to plot the individual positive controls across all samples as well as the slope of
                a linear fit of counts versus input for positive controls in a NanoString ExpressionSet object
}
\author{
Dorothee Nickles
}
\seealso{
\code{\link{posSlopePlot}}
}

