% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{posMaxPlot}
\alias{posMaxPlot}
\title{posMaxPlot}
\usage{
posMaxPlot(rccSet, method = c("cutoffByMMAD", "cutoffByVar"), cutoff)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{method}{Scalar character, specifying the method for outlier detection; either cutoffByMMAD or cutoffByVar}

\item{cutoff}{Scalar numeric, cutoff in method to determine outliers}
}
\value{
A plot
}
\description{
Plot the maximum count of the positive controls for each sample in a NanoString ExpressionSet object.
}
\details{
The maximum count of the positive controls for each sample in the ExpressionSet are plotted and and outliers (as determined
by method) are marked in red (thresholds for outlier definition are plotted as red dashed lines).
}
\author{
Dorothee Nickles
}

