% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{posNormFactPlot,RccSet-method}
\alias{posNormFactPlot,RccSet-method}
\alias{posNormFactPlot}
\title{posNormFactPlot}
\usage{
\S4method{posNormFactPlot}{RccSet}(rccSet)
}
\arguments{
\item{rccSet}{An RccSet object}
}
\value{
A plot of the positive control scaling factor for each sample in an RccSet object.
Samples with a positive control scaling factor < 0.3 or > 3 (thresholds defined by NanoString) are
marked in red (dashed red line indicates threshold).
}
\description{
Plot positive control scaling factor for each sample in an RccSet object
}
\author{
Dorothee Nickles
}
