% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{posR2Plot,RccSet-method}
\alias{posR2Plot}
\alias{posR2Plot,RccSet-method}
\title{posR2Plot}
\usage{
\S4method{posR2Plot}{RccSet}(rccSet)
}
\arguments{
\item{rccSet}{RccSet object}
}
\value{
A plot
}
\description{
Plot the R squared of linear fit of counts versus input for positive controls in an RccSet object.
}
\details{
R squared for each sample in the RccSet are plotted and samples with R squared < 0.95 are marked in red
(threshold indicated by dashed red line).
}
\author{
Dorothee Nickles
}

