% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{posRatioPlot,RccSet-method}
\alias{posRatioPlot}
\alias{posRatioPlot,RccSet-method}
\title{posRatioPlot}
\usage{
\S4method{posRatioPlot}{RccSet}(rccSet, method = c("cutoffByMMAD",
  "cutoffByVar"), stringency = 4)
}
\arguments{
\item{rccSet}{An RccSet object}

\item{method}{Character string specifying the method for outlier detection: either
"cutoffByMMAD" or "cutoffByVar".}

\item{stringency}{Numeric value passed to the cutoff function specified by the method
argument (see the 'd' argument of cutoffByMMAD and cutoffByVar).}
}
\value{
A plot
}
\description{
Plot the ratio of the mean of positive control counts for each sample and
the overall mean of positive control counts in an RccSet object.
}
\details{
The ratio for each sample in the RccSet is plotted and and outliers (as
determined the cutoff function specified by the method argument) are marked
in red (thresholds for outlier definition are plotted as red dashed lines).
}
\author{
Dorothee Nickles
}

