% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{posSlopePlot,RccSet-method}
\alias{posSlopePlot}
\alias{posSlopePlot,RccSet-method}
\title{posSlopePlot}
\usage{
\S4method{posSlopePlot}{RccSet}(rccSet, method = c("cutoffByMMAD",
  "cutoffByVar"), stringency = 4)
}
\arguments{
\item{rccSet}{An RccSet object}

\item{method}{Character string specifying the method for outlier detection: either
"cutoffByMMAD" or "cutoffByVar".}

\item{stringency}{Numeric value passed to the cutoff function specified by the method
argument (see the 'd' argument of cutoffByMMAD and cutoffByVar).}
}
\value{
A plot
}
\description{
Plot the slope of linear fit of counts versus input for positive controls
in an RccSet object.
}
\details{
The slope for each sample in the RccSet are plotted and and outliers (as
determined by the function specified by the method argument) are marked in
red (thresholds for outlier definition are plotted as red dashed lines).
}
\author{
Dorothee Nickles
}

