% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{posSumVsAllSumPlot,RccSet-method}
\alias{posSumVsAllSumPlot}
\alias{posSumVsAllSumPlot,RccSet-method}
\title{posSumVsAllSumPlot}
\usage{
\S4method{posSumVsAllSumPlot}{RccSet}(rccSet, method = c("cutoffByMMAD",
  "cutoffByVar"), stringency = 4)
}
\arguments{
\item{rccSet}{An RccSet object}

\item{method}{Character string specifying the method for outlier detection: either
"cutoffByMMAD" or "cutoffByVar".}

\item{stringency}{Numeric value passed to the cutoff function specified by the method
argument (see the 'd' argument of cutoffByMMAD and cutoffByVar).
(If the median ratio is less than 1, three times this value will be
used.)}
}
\value{
A plot
}
\description{
Plot the ratio of sums of positive control counts to all counts for all
samples in an RccSet object.
}
\details{
The ratio for each sample in the RccSet is plotted and and outliers (as
determined by the cutoff function specified by the method argument) are
marked in red (thresholds for outlier definition are plotted as red
dashed lines).
}
\author{
Dorothee Nickles
}

