% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{preprocRccSet}
\alias{preprocRccSet}
\title{Add background correction and normalization to a NanoString ExpressionSet}
\usage{
preprocRccSet(rccSet, method = c("median", "housekeeping"), hkgenes = NULL,
  hkfeatures = NULL)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object to be used as input.}

\item{method}{Normalization method to use: "median" or "housekeeping".}

\item{hkgenes}{Optional character vector with gene symbols
to be used for normalization if method="housekeeping"
instead of the panel housekeeping features
defined in the input (i.e. those features with
featureData CodeClass == "Housekeeping"). If specified,
all features that match any of the specified symbols will
be used. (To specify specific features, use the hkfeatures
argument instead; see below.)}

\item{hkfeatures}{Optional character vector with full feature names
                        ("<CodeClass>_<GeneName>_<Accession>", e.g.
                        "Endogenous_ACTG1_NM_001614.1") to be used for
                        normalization if method="housekeeping" instead of the
                        panel housekeeping features defined in the input.
                        (Note: if this argument is specified at the same time as
                        hkgenes, an error will be thrown.)}
}
\value{
A copy of the input ExpressionSet with additional matrices in the
assayData for each successive preprocessing step (positive control normalization,
background estimation, background correction, and content normalization).
The final, fully preprocessed data is held in assayData(rccSet)$normData.
Details for each step are stored in the experimentData@preprocessing
list (accessible through preproc(rccSet)).
}
\description{
This function performs the positive control normalization, background
correction, and content normalization steps recommended for NanoString
datasets. For each step, a matrix is added to the assayData of the
resulting ExpressionSet object, and a string is appended to the
experimentData@preprocessing list (accessible through preproc(rccSet)
where rccSet is an ExpressionSet output by this function). Positive
control scaling factors are recorded in the output's phenoData in a
column named 'PosCtrl', and the matrix used for background subtraction
is stored in the assayData as 'bgEstimates'. If housekeeping normalization
is performed, a column labeled 'is.housekeeping' is added to the fData that
indicates which features were used for it.
}
\details{
For more information on the rationale behind the recommended
preprocessing and normalization steps, see the vignette.
}
\examples{
data(example_rccSet)
hknorm_example_rccSet <- preprocRccSet(rccSet = example_rccSet,
    method = "housekeeping")
}
\author{
Dorothee Nickles, Robert Ziman
}
\references{
\href{http://www.nanostring.com/media/pdf/MAN_nCounter_Gene_Expression_Data_Analysis_Guidelines.pdf}{NanoString nCounter(R) Expression Data Analysis Guide (2012)}
}

