% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/previewPNG.R
\name{previewPNG}
\alias{previewPNG}
\title{Create a preview of a PNG}
\usage{
previewPNG(inputFile, outputFile, width, cropHeight, res = 72)
}
\arguments{
\item{inputFile}{Input PNG filename}

\item{outputFile}{Output PNG filename}

\item{width}{Width (in pixels) for the preview image}

\item{cropHeight}{Height (in pixels) for the preview image (if the rescaled input is larger
than this, it will be cropped)}

\item{res}{Output PNG resolution (passed to the `res' argument of png())}
}
\value{
A resized, vertically-cropped preview version of the input PNG.
}
\description{
Generates a resized, vertically-cropped preview version of the input PNG.
}

