% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{rccDir.to.pdata_fdata_adata}
\alias{rccDir.to.pdata_fdata_adata}
\title{rccDir.to.pdata_fdata_adata}
\usage{
rccDir.to.pdata_fdata_adata(rccDir)
}
\arguments{
\item{rccDir}{Directory containing .RCC files}
}
\value{
A list containing matrices (for pdata and adata) and a data frame (for fdata)
that pdata_fdata_adata.to.rccSet() then transforms into a full ExpessionSet.
}
\description{
First stage of readRccBatch(): produces a list containing matrices (for pdata and adata) and a data frame (for fdata) that
                pdata_fdata_adata.to.rccSet() then transforms into a full ExpressionSet (after some further checks and adjustments).
                See also nSolverCsv.to.pdata_fdata_adata().
}
\author{
Robert Ziman
}

