% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{rccFiles.to.pdata_fdata_adata}
\alias{rccFiles.to.pdata_fdata_adata}
\title{rccFiles.to.pdata_fdata_adata}
\usage{
rccFiles.to.pdata_fdata_adata(rccFiles)
}
\arguments{
\item{rccFiles}{Vector of .RCC paths}
}
\value{
A list containing matrices (for pdata and adata) and a data frame (for fdata)
that pdata_fdata_adata.to.rccSet() then transforms into a full ExpessionSet.
}
\description{
First stage of readRccBatch(): produces a list containing matrices (for pdata and adata) and a data frame (for fdata) that
                pdata_fdata_adata.to.rccSet() then transforms into a full RccSet (after some further checks and adjustments).
                See also nSolverCsv.to.pdata_fdata_adata().
}
\author{
Robert Ziman
}

