% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_RccSet-class.R
\docType{methods}
\name{RccSet}
\alias{RccSet}
\alias{RccSet,ExpressionSet-method}
\alias{RccSet,environment-method}
\alias{RccSet,matrix-method}
\alias{RccSet,missing-method}
\title{RccSet constructor methods}
\usage{
RccSet(obj, ...)

\S4method{RccSet}{ExpressionSet}(obj, ...)

\S4method{RccSet}{environment}(obj, ...)

\S4method{RccSet}{matrix}(obj, ...)

\S4method{RccSet}{missing}(obj, ...)
}
\arguments{
\item{obj}{An object of appropriate class}

\item{...}{Passed to methods.}
}
\value{
A new \code{\linkS4class{RccSet}} object.
}
\description{
Constructor methods for making new \code{\linkS4class{RccSet}} objects.
}
\details{
Arguments accepted by constructors are identical to those for the
\code{\link{ExpressionSet}} constructors.

See \code{\linkS4class{RccSet}} class documentation for examples of
constructor use.

Constructor calls for which mandatory \code{phenoData} or \code{featureData}
columns are missing will successfully create objects that include mandatory
columns, but with \code{NA} values. See \code{\linkS4class{RccSet}}
documentation for a list of mandatory columns.
}

