% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{readRccBatch}
\alias{readRccBatch}
\title{Read RCC files}
\usage{
readRccBatch(rccFiles)
}
\arguments{
\item{rccFiles}{Vector of .RCC file paths}
}
\value{
An RccSet object that has raw counts in assayData, probe information
in fData, and sample annotation in pData.
}
\description{
Reads the contents of all .RCC files from a given directory into a new
RccSet object. Note: this function is not intended for external use. For
that, see newRccSet().
}
\author{
Robert Ziman
}

