% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{readRccCollectorToolExport}
\alias{readRccCollectorToolExport}
\title{Read RCC Collector Tool Export}
\usage{
readRccCollectorToolExport(file)
}
\arguments{
\item{file}{Path to the NSolver .CSV file to be read.}
}
\value{
An RccSet object that has count data in exprs, probe
information in fData and sample annotation in pData.
}
\description{
Reads the contents of a .CSV file generated from the RCC Collector Tool Export feature
of NanoString's nSolver Analysis software into a new RccSet object.
(Note: this function is not intended for external use. For that, see newRccSet().)
}
\details{
See 'details' in the readRccBatch() help page.
}
\author{
Dorothee Nickles, Thomas Sandmann
}

