% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{sampleClustering}
\alias{sampleClustering}
\title{sampleClustering}
\usage{
sampleClustering(rccSet, annCol = NULL, log.transform = FALSE, covar, ...)
}
\arguments{
\item{rccSet}{ExpressionSet}

\item{annCol}{See \code{\link{aheatmap}}}

\item{log.transform}{(currently unused)}

\item{covar}{Covariate (e.g. "SampleType")}

\item{...}{Extra arguments passed to aheatmap() (optional)}
}
\value{
Nothing, a plot is generated on the active device.
}
\description{
Clustering by sample correlation
}

