% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{sampleFlag}
\alias{sampleFlag}
\title{sampleFlag}
\usage{
sampleFlag(rccSet, reference = c("blank", "negatives"), stringency = 2.5)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{reference}{scalar character, reference to determine limit of detection; either "blank" or "negatives"; passed on to lodAssess}

\item{stringency}{factor by which deviation (MAD) of the median of the reference counts is multiplied; passed on to lodAssess}
}
\value{
A numeric vector giving the indices of samples to be flagged
}
\description{
Flag samples in an experiment based on the number of non-detectable genes and sample-by-sample correlations - obsolete
}
\details{
Deprecated! - Samples are flagged if their median Pearson correlation coefficient with all other samples in the experiment is below 0.2
or if the number of detected genes (as determined by lodAssess) is less than half the genes assayed.
}
\author{
Dorothee Nickles
}
\seealso{
\code{\link{lodAssess}}
}

