% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{scatterPair}
\alias{scatterPair}
\title{scatterPair}
\usage{
scatterPair(x, y)
}
\arguments{
\item{x}{integer, x positions}

\item{y}{integer, y positions}
}
\value{
A scatter plot x versus y.
}
\description{
Helper function for a scatter plot inside a pairs plots
}

