% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{sumRatioPlot}
\alias{sumRatioPlot}
\title{sumRatioPlot}
\usage{
sumRatioPlot(rccSet, pair, method = c("cutoffByMMAD", "cutoffByVar"), cutoff)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{pair}{list of length two, specifying the two code classes that are compared to each other}

\item{method}{scalar character, specifying the method for outlier detection; either cutoffByMMAD or cutoffByVar}

\item{cutoff}{scalar numeric, cutoff in method to determine outliers}
}
\value{
A plot
}
\description{
Plot the ratio of counts for tho different code classes for all samples in a NanoString ExpressionSet object.
}
\details{
The ratio of counts is determined the following way: counts for code class == pair[[1]] / counts for code class == pair[[2]].
The ratio for each sample in the ExpressionSet is plotted and and outliers (as determined
by method) are marked in red (thresholds for outlier definition are plotted as red dashed lines).
}
\author{
Dorothee Nickles
}

