% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{validRccSet}
\alias{validRccSet}
\title{Validate a NanoString ExpressionSet}
\usage{
validRccSet(rccSet, stopOnError = TRUE, reportWarnings = TRUE,
  showMessages = FALSE)
}
\arguments{
\item{rccSet}{A NanoString ExpressionSet to be validated.}

\item{stopOnError}{Logical. If TRUE and validity fails, the function generates
an error message. If FALSE and validity fails, the function
returns a character vector with the errors found.}

\item{reportWarnings}{Logical. If TRUE, warnings are reported.}

\item{showMessages}{Logical. If TRUE, notes are shown indicating any optional missing
                        columns and the like.}
}
\value{
Returns TRUE if the input object is an ExpressionSet with NanoString
                        data. If problems are found and stopOnError=TRUE, the function
                        will generate an error message. If problems are found and
                        stopOnError=FALSE, the function will return a character vector with
                        the errors found.
}
\description{
Checks to see if the given ExpressionSet contains all the columns and other
member elements that should be present if the ExpressionSet contains
NanoString mRNA assay data.
}
\examples{
data(example_rccSet)
validRccSet(example_rccSet)
fData(example_rccSet)$CodeClass <- NULL
#validRccSet(example_rccSet)    # returns an error
}
\author{
Robert Ziman
}

