% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{zfacFun}
\alias{zfacFun}
\title{zfacFun}
\usage{
zfacFun(p, n)
}
\arguments{
\item{p}{numeric vector: measurements for the positive controls (or actual measurement)}

\item{n}{numeric vector: measurements for the negative controls}
}
\value{
Scalar numeric: the Z' Factor
}
\description{
Calculate Z' Factor
}
\author{
Dorothee Nickles
}

