% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{addQCFlags}
\alias{addQCFlags}
\title{Add sample QC flags to an rccSet}
\usage{
addQCFlags(rccSet, flagFile)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object to be used as input}

\item{flagFile}{Path to a sample QC flag file as generated by the
                    NanoStringQCPro QC report (see makeQCReport())}
}
\value{
A copy of the input ExpressionSet with columns added to pData from the QC
flag file.
}
\description{
Returns a copy of the input ExpressionSet with columns added to pData
                from the provided sample QC flag annotation file. (That file is
                produced by makeQCReport(); see its help page for more
                details.)
}
\author{
Dorothee Nickles
}

