% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{allSumPlot}
\alias{allSumPlot}
\title{allSumPlot}
\usage{
allSumPlot(rccSet, method = c("cutoffByMMAD", "cutoffByVar"), cutoff)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{method}{scalar character, specifying the method for outlier detection; either cutoffByMMAD or cutoffByVar}

\item{cutoff}{scalar numeric, cutoff in method to determine outliers}
}
\value{
A plot
}
\description{
Plot the sum of all counts (endogenous and housekeeping genes only) for each sample in a NanoString ExpressionSet object.
}
\details{
The sum of counts for each sample in the ExpressionSet are plotted and and outliers (as determined
by method) are marked in red (thresholds for outlier definition are plotted as red dashed lines).
}
\author{
Dorothee Nickles
}

