% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{bdPlot}
\alias{bdPlot}
\title{bdPlot}
\usage{
bdPlot(rccSet)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}
}
\value{
A plot
}
\description{
Plot the binding density of each sample in a NanoString ExpressionSet object.
Samples with a binding density < 0.05 or > 2.25 (thresholds defined by NanoString)
are marked in red (dashed red line indicates threshold).
}
\author{
Dorothee Nickles
}

