% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{colByCovar}
\alias{colByCovar}
\title{colByCovar}
\usage{
colByCovar(pdata, covar)
}
\arguments{
\item{pdata}{pData() of an RccSet object}

\item{covar}{character, colname in the pdata used to stratify (color) data}
}
\value{
A list of length 2, with [["color"]] being a character vector of colors (one color for each level of covar) of length=number of observations
and [["legend"]] providing the levels of covar to map colors to covar
}
\description{
Define colors based on a covariate of an RccSet object
}
\author{
Dorothee Nickles
}

