% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{contentNorm}
\alias{contentNorm}
\title{NanoString count data normalization}
\usage{
contentNorm(rccSet, method = c("median", "mean", "housekeeping"), hk = NULL)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{method}{Normalization method (one of "median", "mean", or
"housekeeping")}

\item{hk}{Logical (TRUE/FALSE) vector defining, for each
                        feature, whether or not it shall be used for
                        housekeeping normalization if method="housekeeping"}
}
\value{
A NanoString ExpressionSet object that has log2-transformed and
normalized count data.
}
\description{
Function to normalize count data given a NanoString ExpressionSet object.
Note that count data is log2-transformed before normalization and *remains*
log2-transformed returned ExpressionSet. If housekeeping normalization is
specified, the housekeeping features must be specified as well (see below)
and this will be recorded in the returned ExpressionSet in a new fData column
named "is.housekeeping".
}
\examples{
data(example_rccSet)

pcnorm_example_rccSet <- posCtrlNorm(example_rccSet)
bg <- getBackground(pcnorm_example_rccSet)
bgcorr_example_rccSet <- subtractBackground(pcnorm_example_rccSet, bg)

gmnorm_example_rccSet <- contentNorm(bgcorr_example_rccSet, method="median")
hknorm_example_rccSet <- contentNorm(bgcorr_example_rccSet, method="housekeeping",
    hk=(tolower(fData(bgcorr_example_rccSet)$CodeClass) == "housekeeping"))
}
\author{
Dorothee Nickles
}

