% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{ctrlsOverviewPlot}
\alias{ctrlsOverviewPlot}
\title{ctrlsOverviewPlot}
\usage{
ctrlsOverviewPlot(rccSet)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}
}
\value{
A plot with two panels, one for the negative controls, one for the positive controls.
}
\description{
Plot individual negative and positive controls across all samples in a NanoString ExpressionSet object.
}
\author{
Dorothee Nickles
}

