% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{ctrlsZprimePlot}
\alias{ctrlsZprimePlot}
\title{ctrlsZprimePlot}
\usage{
ctrlsZprimePlot(rccSet)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}
}
\value{
A plot
}
\description{
Plot distribution of counts and the Z' Factors comparing the nagative controls and
                the three highest input positive controls of a NanoString ExpressionSet object.
}
\author{
Dorothee Nickles
}

