% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{flagSamplesCtrl}
\alias{flagSamplesCtrl}
\title{flagSamplesCtrl}
\usage{
flagSamplesCtrl(rccSet, method = c("cutoffByMMAD", "cutoffByVar"), cutoff)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{method}{scalar character, specifying the method for outlier detection; either cutoffByMMAD or cutoffByVar}

\item{cutoff}{scalar numeric, cutoff in method to determine outliers}
}
\value{
A numeric vector giving the indices of samples with outlier values in control metrics
}
\description{
Flag samples based on the performance of their controls.
}
\details{
Outliers are determined based on method and cutoff. Samples defined as outliers in the following categories
will be flagged: interquartile range of negative and positive controls, slope of the linear fit of count versus input
of positive controls, a positive control scaling factor (based on sum of positive controls) below 0.3 or greater than 3.
}
\author{
Dorothee Nickles
}

