% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{fovPlot,RccSet-method}
\alias{fovPlot}
\alias{fovPlot,RccSet-method}
\title{Fields of view (FOV) plot}
\usage{
\S4method{fovPlot}{RccSet}(rccSet)
}
\arguments{
\item{rccSet}{An RccSet object}
}
\value{
A plot
}
\description{
Plot the fraction of successfully imaged fields of view (FOV) in the given RccSet.
The RccSet's phenoData should have 'FovCount' and 'FovCounted' columns populated with the total
and successfully imaged FOV counts, respectively. Samples with a FOV counted/FOV count of less than
80% are marked in red (dashed red line indicates threshold).
}
\author{
Dorothee Nickles
}

