% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BackgroundSubtractionFunctions.R
\docType{methods}
\name{getBackground,RccSet-method}
\alias{getBackground}
\alias{getBackground,RccSet-method}
\title{Get background estimates for a NanoString RccSet}
\usage{
\S4method{getBackground}{RccSet}(rccSet, reference = c("blank", "Negative",
  "both"), metric = c("median", "mean"), stringency = 0,
  nSolverBackground.shrink = TRUE, nSolverBackground.w1 = 2.18)
}
\arguments{
\item{rccSet}{NanoString RccSet object.}

\item{reference}{Measurements to use for background estimates: either "blank" (for blank
samples), "Negative" (for negative control probes), or "both". Blanks are
assumed to be indicated as in the description above.}

\item{metric}{Summarization metric for background measurements: either "mean" or "median".}

\item{stringency}{Factor by which deviation (SD or MAD) of the summarization metric will be
multiplied to obtain final background estimates.}

\item{nSolverBackground.shrink}{Value to use for the 'shrink' argument to nSolverBackground().}

\item{nSolverBackground.w1}{Value to use for the 'w1' argument to nSolverBackground().}
}
\value{
A matrix containing background estimates for a NanoString RccSet object.
}
\description{
Returns background estimates for a NanoString RccSet object. The function
depends upon correct annotation in the RccSet: if the \code{reference}
argument is set to \code{"blank"}, it expects blank measurements (i.e.,
water runs) to have their phenoData SampleType set to the value indicating
blanks (see getBlankLabel(); normally this value would have been set using
an argument to newRccSet()). If \code{reference} is set to
\code{"Negative"}, then it expects to find the negative control probes via
CodeClass == "Negative". If set to "both", it expects both of the above and
will calculate initial background estimates using an algorithm that mimics
the implementation in NanoString's nSolver Analysis Software; see the
nSolverBackground() man page for details on the algorithm.
}
\examples{
data(example_rccSet)

## Calculate probe-specific background based on negative control probes
bg <- getBackground(example_rccSet, reference="Negative", metric="mean")

## Calculate sample-specific background based on blanks
bg <- getBackground(example_rccSet, reference="blank")

## Calculate background that is both sample- and probe-specific
bg <- getBackground(example_rccSet, reference="both", stringency=1)
}
\author{
Dorothee Nickles
}
\seealso{
\code{\link{subtractBackground}}
}

