% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{lodAssess,RccSet-method}
\alias{lodAssess}
\alias{lodAssess,RccSet-method}
\title{lodAssess}
\usage{
\S4method{lodAssess}{RccSet}(rccSet)
}
\arguments{
\item{rccSet}{An RccSet object}
}
\value{
A numeric vector giving the number of missing genes (endogenous and
housekeeping genes) for each sample in an RccSet. If paData is not
found in the input's assayData, NULL is returned.
}
\description{
Assess how many genes in each sample in an RccSet object are below the
limit of detection. (The current implementation does a straightforward
column sum on the presence/absence matrix (paData) in assayData.)
}
\author{
Dorothee Nickles
}

