% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{lodPlot}
\alias{lodPlot}
\title{lodPlot}
\usage{
lodPlot(rccSet, reference = c("blank", "negatives"), stringency = 2.5,
  maxMiss)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{reference}{scalar character, reference to determine limit of detection; either "blank" or "negatives";
passed on to \code{\link{lodAssess}}}

\item{stringency}{factor by which deviation (MAD) of the median of the reference counts is multiplied;
passed on to \code{\link{lodAssess}}}

\item{maxMiss}{scalar numeric, indicating the fraction of genes accepted to be missed}
}
\value{
A plot
}
\description{
Function to plot the number of missing genes per sample in a NanoString ExpressionSet object
}
\details{
Samples with more than 50% missingness are marked in red. If blank measurements are present, they are represented as triangles.
}
\author{
Dorothee Nickles
}

