% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQCreport.R
\name{makeQCReport}
\alias{makeQCReport}
\title{Make NanoString QC report}
\usage{
makeQCReport(rccSet, outputBaseName = "NanoStringQCPro_QC_report",
  outputDir = getwd(),
  experimentTitle = expinfo(experimentData(rccSet))["title"],
  covar = "SampleType", reference = c("auto", "blank", "negatives"),
  method = c("cutoffByMMAD", "cutoffByVar"), cutoff = 4, maxMiss = 0.2,
  SampleName = "SampleID", cleanMarkdown = TRUE)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object for which to generate the QC report}

\item{outputBaseName}{Scalar character, the base (without extension) to use for various output filenames}

\item{outputDir}{Scalar character, the path to the output directory for the QC report and associated files}

\item{experimentTitle}{Scalar character, an easy to read identifier of the experiment}

\item{covar}{Scalar character, covariate for stratifying samples}

\item{reference}{Scalar character, reference to determine limit of detection: "blank" specifies that blank samples
should be used; "negatives" specifies that negative control
probes should be used; "auto" specifies that the decision should be automatic (blank samples will be
used if present, otherwise negative control probes will be used).}

\item{method}{Scalar character, method to determine outliers; either "cutoffByVar" or "cutoffByMMAD"}

\item{cutoff}{Scalar numeric, cutoff in method to determine outliers}

\item{maxMiss}{Scalar numeric, indicating the fraction of genes accepted to be missed}

\item{SampleName}{Scalar character, name of the phenoData column holding the sample names}

\item{cleanMarkdown}{Should markdown files used to produce QC report be deleted upon completion?}
}
\value{
An html report is written to disk, and a copy of the input ExpressionSet is invisibly
returned, with columns added to phenoData that show the QC flags for each sample.
}
\description{
Creates an html QC report for a NanoString ExpressionSet object. Since the report depends upon
correct NanoString-specific annotation in the input ExpressionSet, this function calls
validRccSet() to check for any missing columns or other elements. If any problems are found,
the function will abort with an error that describes them. Otherwise, the function will return
a copy of the input ExpressionSet with additional columns in phenoData showing the QC flags for
each sample.
}
\examples{
data(example_rccSet)

pcnorm_example_rccSet <- posCtrlNorm(example_rccSet)
bg <- getBackground(pcnorm_example_rccSet)
bgcorr_example_rccSet <- subtractBackground(pcnorm_example_rccSet, bg)

qc_example_rccSet <- makeQCReport(bgcorr_example_rccSet, "QC_report")
}
\author{
Dorothee Nickles, Thomas Sandmann, Robert Ziman
}

