% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{negCtrlsPairs}
\alias{negCtrlsPairs}
\title{negCtrlsPairs}
\usage{
negCtrlsPairs(rccSet, log.transform = FALSE)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{log.transform}{boolean, whether data needs to be log2 transformed}
}
\value{
Pairs plot of the negative controls with a scatter plot in the lower panel and
correlatation coefficients printed in the upper panel.
}
\description{
Pairs plot of negative controls across all samples in a NanoString ExpressionSet object
}
\author{
Dorothee Nickles
}

