% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{pcaPlot}
\alias{pcaPlot}
\title{pcaPlot}
\usage{
pcaPlot(exx, ...)
}
\arguments{
\item{exx}{exprs() of a NanoString ExpressionSet object (exprs(rccSet))}

\item{...}{additional parameters passed on to the plotting functions}
}
\value{
PCA screeplot and a plot with two panels, one plotting PC1 versus PC2, the other plotting
PC1 versus PC3.
}
\description{
Wrapper function to perform a PCA analysis on the exprs slot of an ExpressionSet object
                and plot some results
}
\author{
Dorothee Nickles
}

