% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{posCtrlNorm}
\alias{posCtrlNorm}
\title{Positive control normalization}
\usage{
posCtrlNorm(rccSet, metric = c("sum", "mean", "median"),
  recordPosFactor = TRUE)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{metric}{Metric of positive controls used for normalization; one of "mean", "median", or "sum" (the default)}

\item{recordPosFactor}{Logical. If TRUE, 'PosFactor' will be added to the output's phenoData to record the positive control
                        scaling factor that was computed for each sample.}
}
\value{
A NanoString ExpressionSet object that has count data adjusted by positive control counts.
The positive control scaling factor is recorded in PosFactor in the output's phenoData
(an error is generated if this column already exists in the input).
}
\description{
Applies NanoString's recommended positive control normalization to data in a NanoString ExpressionSet object.
}
\examples{
data(example_rccSet)
pcnorm_example_rccSet <- posCtrlNorm(example_rccSet)
}
\author{
Dorothee Nickles
}

