% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{posNormFactPlot}
\alias{posNormFactPlot}
\title{posNormFactPlot}
\usage{
posNormFactPlot(rccSet)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}
}
\value{
A plot of the positive control scaling factor for each sample in a NanoString ExpressionSet object.
Samples with a positive control scaling factor < 0.3 or > 3 (thresholds defined by NanoString) are
marked in red (dashed red line indicates threshold).
}
\description{
Plot positive control scaling factor for each sample in a NanoString ExpressionSet object
}
\author{
Dorothee Nickles
}

