% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{posR2Plot}
\alias{posR2Plot}
\title{posR2Plot}
\usage{
posR2Plot(rccSet)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}
}
\value{
A plot
}
\description{
Plot the R squared of linear fit of counts versus input for positive controls in a NanoString ExpressionSet object.
}
\details{
R squared for each sample in the ExpressionSet are plotted and samples with R squared < 0.95 are marked in red
(threshold indicated by dashed red line).
}
\author{
Dorothee Nickles
}

