% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\name{posSlopePlot}
\alias{posSlopePlot}
\title{posSlopePlot}
\usage{
posSlopePlot(rccSet, method = c("cutoffByMMAD", "cutoffByVar"), cutoff)
}
\arguments{
\item{rccSet}{NanoString ExpressionSet object}

\item{method}{Scalar character specifying the method for outlier detection; either cutoffByMMAD or cutoffByVar}

\item{cutoff}{Scalar numeric: cutoff in method to determine outliers}
}
\value{
A plot
}
\description{
Plot the slope of linear fit of counts versus input for positive controls in a NanoString ExpressionSet object.
}
\details{
The slope for each sample in the ExpressionSet are plotted and and outliers (as determined
by method) are marked in red (thresholds for outlier definition are plotted as red dashed lines).
}
\author{
Dorothee Nickles
}

