% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{readRcc}
\alias{readRcc}
\title{Read an .RCC file}
\usage{
readRcc(rccPath, removeSpikeInLabels = TRUE)
}
\arguments{
\item{rccPath}{Path to the .RCC file.}

\item{removeSpikeInLabels}{Logical. If TRUE (the default), RNA ``spike-in'' input labels (if any)
                            in the GeneName for positive and negative control probes will be removed.}
}
\value{
A list where each element holds the contents of one part of the .RCC file
(Header, Sample_Attributes, Lane_Attributes, Code_Summary, etc) as a
vector or data frame.
}
\description{
Parse an .RCC file into a list with each part of the file (Header,
                Sample_Attributes, Lane_Attributes, Code_Summary, etc) stored as a vector
                or data frame. (Note: for reading a set of .RCC files together, use
                readRccBatch().)
}
\examples{
rccPath <- system.file("extdata", "RCC", "20140604_C1-unstim_C1-unstim_01.RCC", package="NanoStringQCPro")
rcc <- readRcc(rccPath)
}
\author{
Robert Ziman
}

