% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{readRccBatch}
\alias{readRccBatch}
\title{Read a directory of .RCC files to produce an RccSet}
\usage{
readRccBatch(rccDir)
}
\arguments{
\item{rccDir}{Directory containing the .RCC files.}
}
\value{
An RccSet object that has raw counts in assayData, probe information
in fData, and sample annotation in pData.
}
\description{
Reads the contents of all .RCC files from a given directory into a new
RccSet object. Note: this function is not intended for external use. For
that, see newRccSet().
}
\examples{
rccDir <- system.file("extdata", "RCC", package="NanoStringQCPro")
rccSet <- readRccBatch(rccDir)
}
\author{
Robert Ziman
}

