% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanoStringConvenience.R
\name{readRccCollectorToolExport}
\alias{readRccCollectorToolExport}
\title{Read RCC Collector Tool Export}
\usage{
readRccCollectorToolExport(file)
}
\arguments{
\item{file}{Path to the NSolver .csv file to be read.}
}
\value{
A NanoString ExpressionSet object that has count data in exprs, probe
information in fData and sample annotation in pData.
}
\description{
Reads the contents of a .CSV file generated from the RCC Collector Tool Export feature
of NanoString's nSolver Analysis software into a new NanoString ExpressionSet object.
(Note: this function is not intended for external use. For that, see newRccSet().)
}
\details{
See 'details' in the readRccBatch() help page.
}
\examples{
csvPath <- system.file("extdata", "nSolver", "RCC_collector_tool_export.csv", package="NanoStringQCPro")
rccSet <- readRccCollectorToolExport(csvPath)
}
\author{
Dorothee Nickles, Thomas Sandmann
}

