% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQualityMetrics.R
\docType{methods}
\name{sampleClustering,RccSet-method}
\alias{sampleClustering}
\alias{sampleClustering,RccSet-method}
\title{Clustering by sample correlation}
\usage{
\S4method{sampleClustering}{RccSet}(rccSet, outputFile,
  main = "Sample correlations in raw data", annCol = NULL,
  covar = "SampleType")
}
\arguments{
\item{rccSet}{An RccSet}

\item{outputFile}{Output PNG filename}

\item{main}{Plot title}

\item{annCol}{See \code{\link{aheatmap}}}

\item{covar}{Covariate (e.g. "SampleType")}
}
\value{
A PNG file showing clustering of samples by correlation. Any zero-variance
samples are omitted from the heatmap.
The width and height of the PNG file are set to vary with the size of the input.
}
\description{
Clustering by sample correlation
}
\author{
Dorothee Nickles, Robert Ziman
}

