% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BackgroundSubtractionFunctions.R
\docType{methods}
\name{subtractBackground,RccSet-method}
\alias{subtractBackground}
\alias{subtractBackground,RccSet-method}
\title{Subtract background estimates for a NanoString RccSet}
\usage{
\S4method{subtractBackground}{RccSet}(rccSet, bgEstimates, description = "")
}
\arguments{
\item{rccSet}{NanoString RccSet object}

\item{bgEstimates}{Matrix containing the background estimates.}

\item{description}{Description for the background estimates.}
}
\value{
A NanoString \code{linkS4class{RccSet}} object with background estimates
subtracted from the count data.
}
\description{
Returns a NanoString \code{\linkS4class{RccSet}} with background-corrected
count data. During subtraction, any counts below 1 will be truncated to 1 to
enable subsequent log transformation of the data.
}
\examples{
data(example_rccSet)

bg1 <- getBackground(example_rccSet, reference="Negative", metric="mean")
bg2 <- getBackground(example_rccSet, reference="blank")
bg3 <- getBackground(example_rccSet, reference="both", stringency=1)

bgCor1 <- subtractBackground(
    example_rccSet,
    bgEstimates=bg1,
    description="Background estimates using only negative control probes")

bgCor2 <- subtractBackground(
    example_rccSet,
    bgEstimates=bg2,
    description="Background estimates using only blank samples")

bgCor3 <- subtractBackground(
    example_rccSet,
    bgEstimates=bg3,
    description="Background estimates using both blank samples and negative control probes")
}
\author{
Dorothee Nickles
}
\seealso{
\code{\link{getBackground}}
}

